functions/requirements.txt
azure-functions
httpx
azure-storage-blob
azure-core
